Move DependentForm to new "forms" subfolder in "components"
