+++ src/components/forms/DependentForm.svelte
@@ -1,5 +1,5 @@
 <script>
-import RadioOptions from './RadioOptions.svelte'
+import RadioOptions from '../RadioOptions.svelte'
 import { Button, Form, TextField } from '@silintl/ui-components'
 import { createEventDispatcher } from 'svelte'
