+++ src/components/index.js
@@ -7,7 +7,7 @@ import Breadcrumb from './Breadcrumb.svelte'
 import ClaimCard from './ClaimCard.svelte'
 import ClaimCards from './ClaimCards.svelte'
 import DateInput from './DateInput.svelte'
-import DependentForm from './DependentForm.svelte'
+import DependentForm from './forms/DependentForm.svelte'
 import Description from './Description.svelte'
 import MoneyInput from './MoneyInput.svelte'
 import RadioOptions from './RadioOptions.svelte'
